+++ main/options.h
@@ -178,6 +178,7 @@ extern bool processMapOption (const char *const option, const char *const parame
 extern bool processKinddefOption (const char *const option, const char *const parameter);
 extern bool processKindsOption (const char *const option, const char *const parameter);
 extern bool processExtradefOption (const char *const option, const char *const parameter);
+extern bool processFielddefOption (const char *const option, const char *const parameter);
 extern bool processAliasOption (const char *const option, const char *const parameter);
 #ifdef HAVE_ICONV
 extern bool processLanguageEncodingOption (const char *const option, const char *const parameter);